# openapi/aurum.yaml
openapi: 3.0.3
info:
  title: Aurum API
  version: 0.1.0
  description: API for retrieving market intelligence data, managing scenarios, and valuing PPAs.
servers:
  - url: https://api.aurum.local
    description: Local development gateway
  - url: https://api.aurum.prod
    description: Production
security:
  - oauth2: [aurum]
paths:
  /v1/curves:
    get:
      summary: Retrieve curve observations with filtering by identity and tenor.
      operationId: listCurves
      security:
        - oauth2: [aurum]
      parameters:
        - $ref: '#/components/parameters/AsOfDate'
        - $ref: '#/components/parameters/AssetClass'
        - $ref: '#/components/parameters/Iso'
        - $ref: '#/components/parameters/Location'
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Block'
        - $ref: '#/components/parameters/TenorType'
        - $ref: '#/components/parameters/CurveKey'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
          description: Maximum number of rows to return (capped at 500 by the service).
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
          description: Offset for pagination.
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor for stable pagination.
        - name: since_cursor
          in: query
          schema:
            type: string
          description: Alias for `cursor` to resume iteration from a previously returned `next_cursor`.
        - name: prev_cursor
          in: query
          schema:
            type: string
          description: Cursor pointing to the previous page; use the `meta.prev_cursor` value from a prior response.
        - name: format
          in: query
          description: Response format (set to `csv` to stream results).
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: Curve data retrieved.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurveResponse'
            text/csv:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/curves/diff:
    get:
      summary: Diff curve observations between two as-of dates.
      operationId: diffCurves
      security:
        - oauth2: [aurum]
      parameters:
        - name: asof_a
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: asof_b
          in: query
          required: true
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/AssetClass'
        - $ref: '#/components/parameters/Iso'
        - $ref: '#/components/parameters/Location'
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Block'
        - $ref: '#/components/parameters/TenorType'
        - $ref: '#/components/parameters/CurveKey'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Curve diff results.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurveDiffResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /v1/curves/strips:
    get:
      summary: Retrieve non-monthly (strip) curves by type.
      operationId: listStrips
      security:
        - oauth2: [aurum]
      parameters:
        - $ref: '#/components/parameters/AsOfDate'
        - name: type
          in: query
          required: true
          schema:
            type: string
            enum: [CALENDAR, SEASON, QUARTER]
        - $ref: '#/components/parameters/CurveKey'
        - $ref: '#/components/parameters/AssetClass'
        - $ref: '#/components/parameters/Iso'
        - $ref: '#/components/parameters/Location'
        - $ref: '#/components/parameters/Market'
        - $ref: '#/components/parameters/Product'
        - $ref: '#/components/parameters/Block'
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Strip curves retrieved.
          headers:
            ETag:
              description: Strong entity tag for cache validation.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CurveResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    oauth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://auth.aurum.local/oauth2/auth
          tokenUrl: https://auth.aurum.local/oauth2/token
          scopes:
            aurum: Full access to Aurum APIs
  parameters:
    AsOfDate:
      name: asof
      in: query
      required: false
      schema:
        type: string
        format: date
      description: Filter by as-of date.
    AssetClass:
      name: asset_class
      in: query
      schema:
        type: string
      required: false
    Iso:
      name: iso
      in: query
      schema:
        type: string
      required: false
    Location:
      name: location
      in: query
      schema:
        type: string
      required: false
    Market:
      name: market
      in: query
      schema:
        type: string
      required: false
    Product:
      name: product
      in: query
      schema:
        type: string
      required: false
    Block:
      name: block
      in: query
      schema:
        type: string
      required: false
    TenorType:
      name: tenor_type
      in: query
      schema:
        type: string
        enum: [MONTHLY, CALENDAR, SEASON, QUARTER]
      required: false
    CurveKey:
      name: curve_key
      in: query
      schema:
        type: string
      required: false
  responses:
    Unauthorized:
      description: Authentication failed or missing token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    Meta:
      type: object
      properties:
        request_id:
          type: string
        query_time_ms:
          type: integer
        next_cursor:
          type: string
          nullable: true
        prev_cursor:
          type: string
          nullable: true
    CurvePoint:
      type: object
      required: [curve_key, tenor_label, mid]
      properties:
        curve_key:
          type: string
        tenor_label:
          type: string
        tenor_type:
          type: string
        contract_month:
          type: string
          format: date
          nullable: true
        mid:
          type: number
        bid:
          type: number
          nullable: true
        ask:
          type: number
          nullable: true
        price_type:
          type: string
          nullable: true
        asof_date:
          type: string
          format: date
    CurveResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/CurvePoint'
    CurveDiffPoint:
      type: object
      required: [curve_key, tenor_label, asof_a, asof_b]
      properties:
        curve_key:
          type: string
        tenor_label:
          type: string
        tenor_type:
          type: string
          nullable: true
        contract_month:
          type: string
          format: date
          nullable: true
        asof_a:
          type: string
          format: date
        mid_a:
          type: number
          nullable: true
        asof_b:
          type: string
          format: date
        mid_b:
          type: number
          nullable: true
        diff_abs:
          type: number
          nullable: true
        diff_pct:
          type: number
          nullable: true
    CurveDiffResponse:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
        data:
          type: array
          items:
            $ref: '#/components/schemas/CurveDiffPoint'
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        meta:
          $ref: '#/components/schemas/Meta'